Delegate nested transaction management to db clients via the strategy pattern
